src/requirements.txt
tk
pillow
sphinx
sphinx-rtd-theme
pyinstaller